+++ src/main/java/fr/adaming/forum/dao/ICompanyDao.java
@@ -8,9 +8,9 @@ public interface ICompanyDao {
 
 	public Company addCompany(Company company);
 	public Company updateCompany(Company company);
-	public Company deleteCompany(int idCompany);
+	public Company deleteCompany(Long idCompany);
 	
-	public Company getCompanyById(int idCompany);
+	public Company getCompanyById(Long idCompany);
 	public List<Company> getAllCompany();
 	public List<Company> getCompanyByMc(String keyword);
 }
